requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
faster-whisper>=1.0.0
opencc-python-reimplemented>=0.1.7
edge-tts>=6.1.0
python-multipart>=0.0.6
